requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
decouple==0.0.7
Flask==3.1.1
Flask-MySQLdb==2.0.0
Flask-Mail==0.9.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
requests==2.32.4
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
Flask-WTF==1.2.1
PyMySQL>=1.1.0
gunicorn==21.2.0
